reduce page id type to 32 bits

+++ include/tinydb/storage/page.h
@@ -8,7 +8,7 @@ namespace tinydb::storage {
 
 inline constexpr std::size_t PAGE_SIZE = 4096;
 
-using PageId = std::uint64_t;
+using PageId = std::uint32_t;
 using PageBytes = std::array<char, PAGE_SIZE>;
 
 } // namespace tinydb::storage
